pyproject.toml
[tool.poetry]
name = "sales"
version = "0.1.0"
description = "Technical challenge"
authors = ["marcus.santos <mpsoulbass@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
python-environ = "^0.4.54"
Django = "3.2.4"
psycopg2-binary = "^2.9.1"
python-memcached = "^1.59"
djangorestframework = "3.12.2"
drf-yasg2 = "^1.19.4"
PyYAML = "^5.4.1"

[tool.poetry.dev-dependencies]
model-mommy = "^2.0.0"
coverage = "^5.5"
black = "^21.6b0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
